Refactor built-in IdP metadata abstractions
