--- core/src/main/java/org/apache/gravitino/authorization/IdpGroup.java
@@ -1,39 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.apache.gravitino.authorization;
-
-import java.util.List;
-
-/** Built-in IdP group view used by core abstractions. */
-public interface IdpGroup {
-
-  /**
-   * Returns the group name.
-   *
-   * @return the group name
-   */
-  String name();
-
-  /**
-   * Returns the user names in this group.
-   *
-   * @return the user names
-   */
-  List<String> users();
-}
